fix some utils stuff
