+++ src/commands-message/osu/recentlist.ts
@@ -4,7 +4,7 @@ import { playBuilder } from "@builders/plays";
 import { Mode, PlayType } from "@type/osu";
 import { UserType } from "@type/commandArgs";
 import { EmbedBuilderType } from "@type/embedBuilders";
-import { createPaginationActionRow } from "@utils/buttons";
+import { createPaginationActionRow } from "@utils/pagination";
 import { ButtonStateCache } from "@utils/cache";
 import { EmbedType } from "lilybird";
 import type { GuildTextChannel, Message } from "@lilybird/transformers";
