+++ src/embed-builders/background.ts
