+++ src/embed-builders/command.ts
