+++ src/embed-builders/help.ts
