+++ src/utils/constants.ts
@@ -1,3 +1,24 @@
 export const DEFAULT_PREFIX = [";"];
 export const MAX_AMOUNT_OF_PREFIXES = 10;
 export const SPACE = "\0";
+
+export const grades = {
+    SSH: "<:XH_:1057763296717045891>",
+    XH: "<:XH_:1057763296717045891>",
+    SS: "<:X_:1057763294707974215>",
+    SH: "<:SH_:1057763293491642568>",
+    X: "<:X_:1057763293491642568>",
+    S: "<:S_:1057763291998474283>",
+    A: "<:A_:1057763284327080036>",
+    B: "<:B_:1057763286097076405>",
+    C: "<:C_:1057763287565086790>",
+    D: "<:D_:1057763289121173554>",
+    F: "<:F_:1057763290484318360>",
+};
+
+export const rulesets = {
+    osu: "<:osu:1075928459014066286>",
+    mania: "<:mania:1075928451602718771>",
+    taiko: "<:taiko:1075928454651969606>",
+    fruits: "<:ctb:1075928456367444018>",
+};
