+++ src/utils/profile-processor.ts
